3.1 Lesson: Intro to If Statements
1. What do you think the if does to the code under it? Put your answer in a comment in the code.
The if statement is also known as a conditional statement, which allows a computer to make a decision. A program can test for these conditions with the if statement. In this example, if statements do print the message based on the results or number of each different location (robot, ball, goal) in python. If the number of robot locations is smaller than the ball location, it prints out to the user that (“Almost at the ball.”)

2. What is the purpose of indenting in the if statement? How can we tell what code is enclosed in an if branch and what code is not? Answer in a comment.
The purpose of indenting in the if statement is to print out the different results based on the user's answer (value). For example, if b is greater than a, it will print out “a is less than b”, which means true. However, if it’s false, then we can add another if statement to provide the user that if a is greater than b, it will print out “b is less than a.”

3. Change the initial locations of the objects and get the program to output the same thing.

4. What do the operators += and -= do?
The operators += and -= do calculating numbers with plus and minus signs.